ADB's fixes to Generics
