+++ Generics/dumpchan.cpp
@@ -7,7 +7,7 @@ FILE * DumpChan::dfp = stdout;
 void DumpChan::Dump(FILE * fp)
 {
 fprintf(fp,"DumpChan++++++++++++++++++++++++++++++++++++\n");
-fprintf(fp,"Dump file channel = 0x%p\n",dfp);
+fprintf(fp,"Dump file channel = %u\n",(unsigned)dfp);
 fprintf(fp,"DumpChan------------------------------------\n");
 fflush(fp);
 }
