+++ Generics/dumpchan.h
@@ -1,5 +1,3 @@
-#ifndef __DUMPCHAN__H
-#define __DUMPCHAN__H
 #include <stdio.h>
 
 //==============================================================================
@@ -19,4 +17,4 @@ static FILE * dfp;                     // Dumpfile channel
 };
 
 //==============================================================================
-#endif
+
